Clean up and combination
Cleaned up some of the extra (duplicate files) that were created as a
result of changing todo.js to a service. Changed the service to 'return
itself' so that now we can add more methods to the service and use them
in the controllers. I could not get the constant dataURL to work in the
taskService for some reason and also couldn't (didn't spend much time on
it) get the error to return correctly ... will work on that next