Add and Delete task functionality added
Would rather not have to have MainController call newTask.data and just
get the object. But this is a start. Next is complete task functionality
... the control (button) will need to be changed to be easier to tell
the difference between done and not done tasks.

Also created a batch file to start the todue and deployd server. Jamie
you will need to change the path in the file to make it work for you or
setup your db just like mine so it works for both of us.